FAI-1949 - Delete a search alert

+++ src/SFA.DAS.FAA.Api/ApiResponses/GetSavedSearchesResponse.cs
@@ -31,7 +31,7 @@ public record SavedSearchDto(
     public record SearchParametersDto(
         string? SearchTerm,
         List<int>? SelectedRouteIds,
-        decimal? Distance,
+        int? Distance,
         bool DisabilityConfident,
         List<int>? SelectedLevelIds,
         string? Location,
