+++ src/SFA.DAS.FAA.Domain/Models/SearchParameters.cs
@@ -6,7 +6,7 @@ namespace SFA.DAS.FAA.Domain.Models;
 public record SearchParameters(
     string? SearchTerm,
     List<int>? SelectedRouteIds,
-    decimal? Distance,
+    int? Distance,
     bool DisabilityConfident,
     List<int>? SelectedLevelIds,
     string? Location,
